third_party/blink/web_tests/external/wpt/html/cross-origin-embedder-policy/reflection-credentialless.tentative.https.any.worker-expected.txt

This is a testharness.js-based test.
[FAIL] reflection-credentialless
  assert_equals: expected "credentialless" but got "unsafe-none"
Harness: the test ran to completion.
